Revert "feat(functions): implement handle_message function to streamline message processing and response handling"
This reverts commit 9b6167a14155b000384692525736f1e1765e8447.
