+++ apps/telegram-bot/src/functions.rs
@@ -12,8 +12,7 @@ use sea_orm::{
 };
 use services::{MusicLinkInput, MusicLinkService};
 use teloxide::{
-    prelude::*,
-    types::{Message, ParseMode, ReactionType},
+    types::Message,
     utils::html::{link, user_mention},
 };
 
@@ -175,47 +174,5 @@ pub async fn after_process_message(
             };
         to_insert.insert(db).await?;
     }
-    Ok(())
-}
-
-pub async fn handle_message(
-    bot: Bot,
-    msg: Message,
-    db: Arc<DatabaseConnection>,
-) -> Result<(), teloxide::RequestError> {
-    let chat_id = msg.chat.id;
-    let text = msg.text().unwrap_or_default();
-
-    match process_message(text.to_string(), &msg, db.clone()).await {
-        Err(e) => {
-            tracing::error!("Failed to process message: {}", e);
-        }
-        Ok(response) => match response {
-            ProcessMessageResponse::NoUrlDetected => {
-                tracing::debug!("No URL detected in message, ignoring");
-            }
-            ProcessMessageResponse::HasUrlNoMusicLinksFound => {
-                tracing::debug!("URL detected but no music links found, reacting with sad emoji");
-                bot.set_message_reaction(msg.chat.id, msg.id)
-                    .reaction(vec![ReactionType::Emoji {
-                        emoji: "😢".to_string(),
-                    }])
-                    .await?;
-            }
-            ProcessMessageResponse::HasUrlMusicLinksFound {
-                text,
-                music_link_ids,
-            } => {
-                tracing::info!("Sending music link response to chat {}", chat_id);
-                bot.send_message(msg.chat.id, text)
-                    .parse_mode(ParseMode::Html)
-                    .await?;
-                tracing::debug!("Deleting original message");
-                bot.delete_message(msg.chat.id, msg.id).await?;
-                after_process_message(&msg, &db, music_link_ids).await.ok();
-            }
-        },
-    };
-
     Ok(())
 }
