+++ apps/telegram-bot/src/main.rs
@@ -1,10 +1,14 @@
 use std::sync::Arc;
 
+use functions::{ProcessMessageResponse, after_process_message, process_message};
 use schematic::{Config, ConfigLoader, validate::not_empty};
 use sea_orm::{Database, DatabaseConnection};
 use sea_orm_migration::MigratorTrait;
 use serde::Serialize;
-use teloxide::prelude::*;
+use teloxide::{
+    prelude::*,
+    types::{ParseMode, ReactionType},
+};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 mod functions;
@@ -48,7 +52,43 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let handler = Update::filter_message().endpoint(
         |bot: Bot, db: Arc<DatabaseConnection>, msg: Message| async move {
-            functions::handle_message(bot, msg, db).await
+            let chat_id = msg.chat.id;
+            let text = msg.text().unwrap_or_default();
+
+            match process_message(text.to_string(), &msg, db.clone()).await {
+                Err(e) => {
+                    tracing::error!("Failed to process message: {}", e);
+                }
+                Ok(response) => match response {
+                    ProcessMessageResponse::NoUrlDetected => {
+                        tracing::debug!("No URL detected in message, ignoring");
+                    }
+                    ProcessMessageResponse::HasUrlNoMusicLinksFound => {
+                        tracing::debug!(
+                            "URL detected but no music links found, reacting with sad emoji"
+                        );
+                        bot.set_message_reaction(msg.chat.id, msg.id)
+                            .reaction(vec![ReactionType::Emoji {
+                                emoji: "😢".to_string(),
+                            }])
+                            .await?;
+                    }
+                    ProcessMessageResponse::HasUrlMusicLinksFound {
+                        text,
+                        music_link_ids,
+                    } => {
+                        tracing::info!("Sending music link response to chat {}", chat_id);
+                        bot.send_message(msg.chat.id, text)
+                            .parse_mode(ParseMode::Html)
+                            .await?;
+                        tracing::debug!("Deleting original message");
+                        bot.delete_message(msg.chat.id, msg.id).await?;
+                        after_process_message(&msg, &db, music_link_ids).await.ok();
+                    }
+                },
+            };
+
+            respond(())
         },
     );
